Asteroid Destruction Update
Previously, I was trying to add each broken off piece of the asteroids to a node holder, Asteroids, in the world. This caused lots of issues b/c I needed to emit a signal from new instances of asteroids that I was struggling to connect to the world.

Currently, I have dropped that idea and we are now just adding the new asteroids to the current ones. But as I type this I see how this doesn't work. If the asteroid were to break into chunks and the original chunk was to be freed, this would free all chunks attached to it.... ughhh