Player Movement, Health
Player Movement
Movement bug fixed. When playing with a controller and moving and aim with the left stick, forward angle calculation would cause the player to not move at full speed.

Added the beginning of the players health. Player probably need to be turned in to a State Machine, but I cant' decide what the states should be. Player tends to be in the same state while moving and attacking. But I think a dead state would be helpful, and now that I'm typing it out a state for boosting or "specials" could be nice.

Enemy needs a dead state.