Make Enemy AI
Want to move lots of the enemy script into an AI script that can be added to other enemy types. This AI scene should hold all the Area2Ds and the FiniteStateMachine.

Enemy Script should just hold reference to the AI just like the Health, and Weapon.

Move shoot_laser signal into a global? This does seem convenient. But there maybe a better way? Still should advance with this, refractor later if it seems like a problem.