Enemy & AI
OH BABY,

Enemy's FSM has been updated to work as it's own AI node. This will allow us to attach it to other scenes that are not just enemies.

The movement for the AI is getting REAL noice. Very smooth. Movement works very much like the players. AI can also now wander when the player isn't within it's colliders.

TO-DO:
The AI is no longer just for the enemy, but is still named with the idea that it is.  Will need to rename it's functions names, scene names, signal names, etc...

EnemyChase to AIChase.
enemy_movement to ai_movement