Created an Ally
Ally works sorta, still need to iron out some of the code as it was just copied from the enemy. Probably need to rename the script to something more fitting than "enemy".

Probablly need to look at how the youtuber handled collisions? He seemed to just drop the ally in and it worked, but I know I need to edit all the collisions. Also need to switch how the bullets get initialized. Currently using the group of the parent to set it up, but now that was have a team node its probably better to use.

We can use the team not for more things. IE the player picking up an item that provides shield to allies, having the teams set up via the team node will make this easier than us trying to figure out the grouping.