Bases Done-ish, Updating Shield
Base are doneish. There is still a bug, that's causes a base not to be capturable. Looks like it registers the bases as contested even when enemies are gone.

Updating the shield to come back after it breaks. It works, but is a little janky. I'm thinking the fading in and out of the sprite, might be better done as an animation?