Context Based Movement!
Context base movement has been implemented in our Actor class. Testing is still needed!

This allows the actors to check how close an obstacle is to them and avoid it.

This helps solve our current problem of actors stacking up on each other. Not only was this a bit rudimentary, it's also caused unforeseen consequences such as  when a stack of actors would end up on top of each other, their shields would protect every actor under it.

This will also be helpful when we add some sort of Pathfinding as well as more concrete objects into the game, Asteroids, Space Debris, Space Stations, Planets, etc...  Things we wont want the player or units to be over.