Updated Actor Movement & Weapons Manager
AttackStanding State is no longer in use. Enemies now try to keep a steady distance from their target as long as it's in the attackable group. Might be able to use AS State to have the actors come to a move complete stop, currently they will sometimes bobble back and forth.

Updating Weapons to use a Weapon Manager. This will allow us to change weapons more easily, as well as give variety to our enemies, and allies.